+++ tests/effects_tests/test_blackhole.py
@@ -11,7 +11,7 @@ from terminaltexteffects.utils.graphics import Color, Gradient
 
 @pytest.mark.parametrize(
     "input_data",
-    ["empty", "single_char", "single_column", "single_row", "medium", "tabs"],
+    ["single_char", "single_column", "single_row", "medium", "tabs"],
     indirect=True,
 )
 def test_blackhole_effect(input_data: str, terminal_config_default_no_framerate: TerminalConfig) -> None: