+++ tests/effects_tests/test_bubbles.py
@@ -14,7 +14,7 @@ def _make_terminal_config(existing_color_handling: str) -> TerminalConfig:
 
 @pytest.mark.parametrize(
     "input_data",
-    ["empty", "single_char", "single_column", "single_row", "medium", "tabs"],
+    ["single_char", "single_column", "single_row", "medium", "tabs"],
     indirect=True,
 )
 def test_bubbles_effect(input_data, terminal_config_default_no_framerate) -> None: