+++ tests/effects_tests/test_burn.py
@@ -21,7 +21,7 @@ def _make_terminal_config(existing_color_handling: Literal["always", "dynamic", 
 
 @pytest.mark.parametrize(
     "input_data",
-    ["empty", "single_char", "single_column", "single_row", "medium", "tabs"],
+    ["single_char", "single_column", "single_row", "medium", "tabs"],
     indirect=True,
 )
 def test_burn_effect(input_data: str, terminal_config_default_no_framerate: TerminalConfig) -> None: