+++ tests/effects_tests/test_rings.py
@@ -30,7 +30,7 @@ def _get_ring_character(iterator: effect_rings.RingsIterator) -> effect_rings.Ef
 
 @pytest.mark.parametrize(
     "input_data",
-    ["empty", "single_char", "single_column", "single_row", "medium", "tabs"],
+    ["single_char", "single_column", "single_row", "medium", "tabs"],
     indirect=True,
 )
 def test_rings_effect(input_data: str, terminal_config_default_no_framerate: TerminalConfig) -> None: