+++ tests/effects_tests/test_smoke.py
@@ -30,7 +30,7 @@ def _get_inactive_character(iterator: effect_smoke.SmokeIterator) -> effect_smok
 
 @pytest.mark.parametrize(
     "input_data",
-    ["empty", "single_char", "single_column", "single_row", "medium", "tabs"],
+    ["single_char", "single_column", "single_row", "medium", "tabs"],
     indirect=True,
 )
 def test_smoke_effect(input_data: str, terminal_config_default_no_framerate: TerminalConfig) -> None: