+++ tests/effects_tests/test_spotlights.py
@@ -39,7 +39,7 @@ def _place_spotlight_on_character(
 
 @pytest.mark.parametrize(
     "input_data",
-    ["empty", "single_char", "single_column", "single_row", "medium", "tabs"],
+    ["single_char", "single_column", "single_row", "medium", "tabs"],
     indirect=True,
 )
 def test_spotlights_effect(input_data: str, terminal_config_default_no_framerate: TerminalConfig) -> None: